Messages: add format(Person) for formatting a Person for display
Commands may sometimes print a user-friendly representation of a person.
For example, when adding a person, we get:

    New person added: John Doe Phone: 12345 Email: john@doe.com Address: Dough Ave Tags:

The user-friendly representation (`John Doe Phone: 12345 Email...`) is
formatted using Person#toString(). This is not very ideal because it
means that the model is dealing with user-interface-related concerns,
which it should not be doing because that is the responsibility of the
logic and ui layers. Furthermore, using the toString() method for this
task means that Person#toString() can't be used for more useful
endeavours such as providing information for debugging purposes.

As such, let's introduce Messages#format(Person) which would be a common
method for formmating a Person for display to the user, and teach
commands that need that functionality to use it.

Co-authored-by: Zhaoqi <e0725180@u.nus.edu>