LO-CouplingCohesion: update to match the current codebase
The LO references the `Formatter` class, which does not even exist in
the current codebase at all.

Instead, let's replace it with a completely new example of applying the
single responsibility principle in the current codebase -- the use of a
separate ParserUtil class for handling the parsing of model data.

While we are here, let's add some explanatory comments to make it clear
what the division of responsibilities between the model/parser layers
are.