LO-DIP: update to match the current codebase
In the current codebase, the `LogicManager` class already depends on the
interface `Storage`, so the exercise is not applicable any more.

As such, let's take the easier route to update this LO to match the
current codebase -- we remove the exercise, and replace it with a
statement on how the current design is an application of DIP.

While we are here, update the colors used in the diagram to match the
class diagrams in the developer guide, and rename the diagram so that
its filename better matches its contents.