LO-DI: update to match the current codebase
The LO refers to changing the implementation of the `Logic` (now
`LogicManager`) class so that it depends on the abstract class `Storage`
instead of the concrete class `StorageFile` (now `StorageManager`).

However, in the current codebase, the `LogicManager` class already
depends on the interface `Storage`, so students would not need to do the
above anymore. Update the LO and its diagrams to take this into account.

The other part of the exercise (implementing a `StorageStub` class) is
still applicable.

While we are updating the diagrams, also update the colors used so that
it matches the class diagrams in the developer guide.