Developer Guide: Specify the label for inter-document xref
Linking between different documents produces different behavior compared
to linking within a document, resulting in links being rendered with
the document name as the body instead of the section name.

In our case, links are being rendered with the text DeveloperGuide.html
which does not accurately reflect the destination of the link.

Let's explicitly specify the link text instead.