Tutorial: Removing field
In lieu of a tutorial on how to morph which can be hard to write and maintain,
we provide a simpler and shorter tutorial on how to remove a field instead.

Students trying to morph the product can repeat this tutorial to get the model
to a state where they can work with.