Prevent HelpWindow from maximizing
Address a regression in ebbb713 where subsequent `help` commands will
spawn a `HelpWindow` maximized, covering the screen.

Disable the resizability and the maximize button completely as it serves
no purpose.

Remove unnecessary call to sizeToScene() as HelpWindow defaults to 
USE_COMPUTED_SIZE and the computed size of all its children nodes are 
available before it is shown.

Remove TODO in HelpWindow.fxml as the initial size is set to computed
size.