Add Tutorials to header in documentation
Without a direct link to the tutorials, students find it hard to access
them. To improve ease of access to the tutorials, we add a new
Tutorials category to the navigation bar in the document.

Also, we shorten the site-name to 'AB-3' to maintain the length of the navbar
on common viewports.