Update checking for Duplicates in EditCommand
Editing a Person in certain ways allows duplication of Persons in
the addressbook.

The current implementation does not check for all cases correctly.

Let's use a separate Model with only the relevant Persons to check
for Duplicates.

The check for duplicate Persons comes from the isSamePerson method in
Person class which checks for the same name while having at least one
other identity field(Phone or Email) being the same. To reuse the
checking mechanism from the Model, a simple way would be to remove the
personToEdit from a model and have a separate check.

There are other ways such as modifying the filtered list for checking,
however we need to introduce methods which does not make use of the
existing Model interface.