Update checking for Duplicates in EditCommand
Editing a Person in certain ways allows duplication of Persons in
the addressbook.

The current implementation does not check for all cases correctly.

Let's
* remove the target personToEdit from the model
* add the Person back to its original position in the filtered list
after checking
* extend the functionality of the Model to add a person at a
specified index

To reuse the checking mechanism from the Model, a simple way would be
to remove the personToEdit from the model to check with only the
relevant Persons.

The UniquePersonList in AddressBook currently does not support adding a
Person at a specified position, however, it is needed for the checking
as we have to add back the deleted personToEdit to restore the exact
internal list.