Navigation bar is not necessary for the pdf pages
Adding a css print query will allow for the pages to be
printed out without showing the navigation bar

Let's specify the behaviour of the media print query
to change the display type of the wrapper div inside the site-header
div, which contains the navigation bar, to none. Let's also add a generic
AB-3 title in the css before pseudo selector to improve the display of
the print pdf version.

We found this fix to be the most suitable as
it only modifies the body content when printing the document