Omit the top navigation menu when a page is printed
When a page in the product website (e.g, the User Guide) is converted
into a print view, the top navigation bar appears in printed version.
This is not appropriate as a navigation bar belongs in a Web page but
not in a printed view. This is particularly a problem when documents
in the website are converted into PDF, which is done by 'printing' the
page into a PDF file.

Let's update the behaviour of the media print query to replace the top
navigation menu with simply the project name.