Font Issue on M1 Mac Big Sur
The font used in AB3 (Segoe UI) is not preinstalled for M1 Mac Big Sur,
causing the text in AB3 to be illegible.

Lets change the default font to Open Sans, as it has the 3 font styles
used in AB3. Lets also embed this font to ensure all OS can access the
font.