Upgrade JavaFX to version 16
The FXML files are a mix of JavaFX 8 and 11, which are loaded by JavaFX
11 without any issues. Modifying the files using Scene Builder that
utilises newer versions of JavaFX largely works, however, there are
warning messages about loading a newer version within JavaFX 11.

It is not easy to download Scene Builder that specifically saves FXML as
JavaFX 11. Hence, it is likely that any modifications using Scene
Builder will cause warnings to appear. While it is possible to manually
override version within the FXML files, inconsistency tends appear given
that not everyone might be aware of the issue. This might also result in
repeated commits trying to correct the issue.

The proposal is to upgrade JavaFX to its latest version, alongside all
the FXML files.

The latest version of JavaFX will be compatible with previous versions
and avoid the warnings. By updating the FXML files, consistency can be
achieved.