Upgrade Codecov Action
Codecov GitHub Action v1 uses a deprecated bash uploader, which may
cease to function soon. The v2 version also has a files flag instead of a
file flag, suggesting possible deprecation of the file flag.

In addition, the original filename syntax did not work when testing the
Codecov upgrade on RepoSense for Windows CI. Splitting the filepath
into directory and filename allows Windows CI to work.

Let's upgrade to v2, use the files flag for coverage.xml and the directory
flag for the directory in which coverage.xml is found.