Update Checkstyle checks
There are some checks which enforce the Java basic and intermediate
coding standard as well as the Google Java Style Guide. As an example,
the MethodParamPad check ensures that a method or constructor name is
attached to its open parenthesis. The JavadocTagContinuationIndentation
check is for the Google style guide rule that continuation lines for
block tags are indented 4 or more whitespaces relative to the @.

The default throwsIndent value for the Indentation check is 4
whitespaces although the correct value should be 8.

Let's update the throwsIndent value, add the new checks and fix the
HelpWindow file for the JavadocTagContinuationIndentation check.