Allow CI to pass if Codecov fails
As Codecov service is known to fail intermittently, let's
update gradle.yml to allow CI to pass even if the Codecov
task fails.

Let's also update the DevOps guide to explain this change.
At the same time, let's delegate the Codecov instructions
to a se-edu guide so that that can be maintained more
centrally.