Show warning when ignoring repeated parameters
Repeated occurrences of the same parameter in a command are being
ignored without any notification to the user.
This can lead to potential user errors going unnoticed, especially when
using the wrong parameter prefix multiple times.

Let's introduces a warning system that alerts the user when a
parameter is being ignored.
Now, if a parameter is repeated in a command,
a warning message will be displayed to ensure that
the user is aware of any unintentional mistakes or typos.