Restore functionality to unreachable catch blocks
Classes that implements the methods,
- readAddressBook() from AddressBookStorage interface
- readUserPrefs() from UserPrefsStorage interface
does not actually throw any IOException as such exceptions are caught
within the method and is being treated as DataConversionException
which may not be the case.

All IOException are caught by JsonUtil.readJsonFile() which is called by
classes implementing these methods.

This method also causes ConfigUtil.readConfig() to not throw
IOExceptions when handling actual IOExceptions that is not caused by
a conversion error.

This causes the wrong log messages to be logged when an actual
IOException occur.

Let's
- decouple actual conversion errors from IOExceptions and restore
functionality to the catch blocks in MainApp.java that handles
IOExceptions.
- add catch blocks to handle IOException when reading config files