Improve error message if no write permission
Insufficient permission to write to the data file will show up as
AccessDeniedException in the error message.
The application should be able to provide more informative error message
instructing the user to handle and recover from such errors

Let's
- make the error message more user friendly by informing
the user that there are insufficient permissions for the file
and instruct them to check the relevant file permissions.