Show warning when ignoring repeated parameters
Repeated occurrences of the same parameter in a command are being
ignored without any notification to the user.
This can lead to potential user errors going unnoticed, especially when
using the wrong parameter prefix multiple times.

Let's introduces a warning system that alerts the user when a parameter
is being ignored.
Now, if any parameter is repeated in a command, a warning message ensuring
that the user know which all the parameters are being duplicated.