Fix log level not affecting loggers
Log handlers being set prior to init causes their log level to
remain at INFO rather than updaing to the latest log level.
This causes changing the log level in config to not affect any logging.

As the root logger defaults at INFO, this causes all loggers created to
be at INFO so finer log levels will never be reflected.
As there are many other packages that depends on the root logger,
a separate base logger for the application should be created so that
all loggers created in the application derive from it.
This enables finer control over log levels of logger in the application
without affecting loggers from other packages.

Let's
- Set the log level of handlers in init() to allow log level changes
to be reflected in the log file
- Create a base logger to enable finer log level to be available to
the user