Improve error message if no write permission
Insufficient permission to write to the data file will show up as
AccessDeniedException in the error message.
The application should be able to describe the error to the user rather
than just showing the name of the Exception.

Let's
- make the error message more descriptive by informing
the user that there are insufficient permissions for the file for
AccesssDeniedExceptions.

- Also include a better general purpose description for other forms
of IO related exceptions