Removal of unnecessary catch blocks
Classes that implements the methods,
- readAddressBook() from AddressBookStorage interface
- readUserPrefs() from UserPrefsStorage interface
does not actually throw any IOException as such exceptions are caught
within the method and is being treated as DataConversionException
which may not be the case.

All IOException are caught by JsonUtil.readJsonFile() which is called by
classes implementing these methods.

As the handling of exceptions caused by JSON parsing error and actual
IOExceptions are the same, the current code base can be simplified by
moving the logging of these related errors to within
JsonUtil.readJsonFile() method. This also removes the need for the
abovementioned interfaces to throw IOException as well as the
unutilized catch blocks.

As both IOException and JSON Parsing related exceptions are handled
within readJsonFile() method, the naming of DataConversionException is
no longer appropriate and hence be renamed to ConfigLoadingException
where it is more suited in the current use case.