Removal of repeatable parameters as feature
The behavior of repeated parameters in AddCommand and EditCommand
is to take the last value used.
This can be confusing and hard to work with.
Therefore, instead of allowing repeatable parameters, it is better to
display them as a Command Error to the user.

The checking of duplicate parameter should occur before checking of
validity of individual fields. That is to suggest that suppose an Add or
Edit Command contains multiple duplicate fields that are all invalid,
duplicate field parse error will take precedence over other checking of
individual parameters.