Rename DataConversoinException -> ConfigLoadingException
Classes that implements the methods,
- readAddressBook() from AddressBookStorage interface
- readUserPrefs() from UserPrefsStorage interface
does not actually throw any IOException as such exceptions are caught
within the method and is being treated as DataConversionException
which may not be the case.

All IOException are caught by JsonUtil.readJsonFile() which is called by
classes implementing these methods.

As the handling of JSON parsing error and actual IOExceptions are the
same, the current code base can be simplified by moving the logging of
these related errors to within JsonUtil.readJsonFile() method.

As such, using the naming of DataConversionException is no longer
appropriate and it is better to update it to a more appropriate name.
The relevant logger messages related to the this rename should be updated
as well.