Update ug regarding copying commands
In the pdf version of UG, commands that span across more than one line
when copied-pasted to the application will drop the line-break. As
line-breaks will usually be formatted at spaces when converting to pdf,
this will cause the space in command to be dropped when pasting into
the application.

Therefore, let's update the UG to inform the users about this behavior.