Removal of unnecessary catch blocks and exceptions
In the previous commit, the handling of IOExceptions has been
centralized within the JsonUtil.readJsonFile() method. As a result,
catch blocks and IOExceptions in interfaces such as Storage,
AddressBookStorage, and UserPrefsStorage are no longer necessary.

Additionally, some logger messages need to be updated to reflect this
change.

Therefore, let's remove the unused code and update the relevant log
messages.