Reduce code duplication in build.gradle
Importing dependencies for JavaFX in build.gradle is highly repetitive.
This repetition makes it difficult to read, modify, and edit modules if
necessary. Currently, for each JavaFX module, the dependencies for
'win', 'mac' and 'linux' are individually added. To address this issue,
it would be more efficient to use a nested for loop to handle these
imports.

By implementing a nested for loop, we can reduce code duplication and
improve the ease of modifying modules and platforms. Simplifying the
import process for JavaFX dependencies will enhance overall code
readability and maintainability.

Therefore, let's reduce code duplication by implementing a nested for
loop to handle the import of JavaFX dependencies in build.gradle.