Fix not visible help menu
Launching the application on Linux distributions using KDE as the
desktop manager resulted in the window not being visible when no
minimum window size was set. The problem seems to be specific to the
KDE Plasma graphical desktop environment. This has caused the help
menu to be not visible on such platforms.

Manual testing indicates that bumping the JavaFX version to 13 or higher
resolves the issue. Since version 16 will log a warning with the current
project configuration, let's bump the JavaFX version to 15.