Removal of repeatable parameters as a feature
Currently, the behavior of repeated parameters in AddCommand and
EditCommand is to consider the last value used. This behavior can be
confusing and difficult to work with. Therefore, instead of allowing
repeatable parameters, it is better to present them as a Command Error
to the user.

To implement this change, the checking for duplicate parameters should
be performed before checking the validity of individual fields. In other
words, if an Add or Edit Command contains multiple duplicate fields that
are all invalid, the parsing error for duplicate fields will take precedence
over other checks for individual parameters.