Removal of repeatable parameters as a feature
If the same parameter is repeated multiple times in a command, all
except the last occurrence are ignored silently. This can cause certain
user errors to go unnoticed. Therefore, instead of allowing repeatable
parameters, it is better to treat them as an error instead.

Therefore, if any of the single-valued parameters are repeated more than
once, a user error will be displayed to the user.

# Conflicts:
#	src/main/java/seedu/address/commons/core/Messages.java