Move Messages class into the logic package
The Messages class contains common user-facing messages used by
commands and parsers in the Logic component.

However, Messages class is in the seedu.address.commons.core
package that contains classes used system-wide. Given that the
Messages class is user-facing, it would only ever be used by the
logic and ui components, and thus would fit better inside the
seedu.address.logic package.

Furthermore, Messages class accesses classes in the Model component.
This is not right as the Model component is in a higher layer
compared to the commons package, as per how the layered
architecture should work.

Let's move the Messages class to the seedu.address.logic package.