Update UML diagrams to resolve rendering issues
PlantUML(1.2023.2) introduced a change in behavior when rendering
diagrams. This causes some of the diagrams to be rendered incorrectly or
fail to compile. Diagrams are also being rendered in a different theme
due to a change in the default color theme.

Therefore, some of the PlantUML code and images need to be updated to
ensure functionality and consistency.

Let's
 * update PlantUML code to ensure they can be rendered properly
 * update the diagrams to ensure consistency between the PlantUML code
   and the rendered image