Adjust logger level change for consistent log output
Some CONFIG level log messages are not shown to the users despite
setting to the appropriate log levels as the changing of log level
happens after these CONFIG logs.

Therefore, the function responsible for changing the log level is
shifted to an earlier place in the code.

This ensures that logs between the previous position and the new
location are now logged at the appropriate log level. The adjustment
results in consistent log output, improving the clarity and accuracy of
the logged information.