Fix alignment by changing index font
For a non-monospace font such as Segoe UI, "1."
takes up less horizontal space than "2.". This
results in the first name having different
vertical alignment compared to all the names
below the first one in the PersonListCard.

Lets change the font of the index in the
PersonListCard to a monospaced font such as
Arial Mono, which seems to eliminate the
misalignment issue so that the names in the
PersonListCard are aligned.