Update Jackson dependencies to latest version (v2.17.2)
jackson-databind v2.7.0 and jackson-datatype-jsr310 v2.7.4 are severely
outdated, and have critical security vulnerabilities. However, newer
versions of the library serialize relative paths as absolute, causing
tests to fail when updating the dependencies. This is fixed by modifying
the ObjectMapper to use ToStringSerializer for the Path class, which
serializes relative paths correctly.

Let's update the Jackson dependencies to the latest version (v2.17.2) to
resolve the security vulnerabilities and remove the IntelliJ warning.