UserGuide.md: Make JAR download step generic (#28)
Currently, every time a new release is done, developers would have
to update the file name to reflect the latest JAR file published. 
This can be difficult to keep track of.

Let's change the JAR download step (step 2) in the Quick Start
section in the user guide to say "latest `.jar` file" instead of 
"latest `addressbook.jar`".

This serves to make the JAR download step generic, eliminating the
need for developers to update the file name every time a new
release is published.